+++ packages/terra-folder-tree/src/FolderTree.jsx
@@ -4,7 +4,7 @@ import classNames from 'classnames/bind';
 
 import ActionHeader from 'terra-action-header';
 
-import TreeItem from './subcomponents/TreeItem';
+import FolderTreeItem from './subcomponents/FolderTreeItem';
 import styles from './FolderTree.module.scss';
 
 const cx = classNames.bind(styles);
@@ -15,34 +15,27 @@ const propTypes = {
    */
   children: PropTypes.node.isRequired,
   /**
-   * The title of the folder tree. TODO: improve this description
+   * The title of the folder tree.
    */
-  title: PropTypes.string,
+  title: PropTypes.string.isRequired,
 };
 
-const FolderTree = ({ children, title }) => {
-  console.log(children);
-  
-  return (
-    <>
-      <ActionHeader
-        id="folder-tree-header" // TODO: may need to update this to use unique IDs
-        text={title}
-        level={3}
-      />
-      <ul
-        className={cx('folder-tree')}
-        role="tree"
-        aria-labelledby="folder-tree-header"
-      >
-        {children}
-      </ul>
-    </>
-  );
-};
+const FolderTree = ({ children, title }) => (
+  <>
+    <ActionHeader
+      text={title}
+      level={3} // does this need to be configurable?
+    />
+    <ul
+      className={cx('folder-tree')}
+      role="tree"
+    >
+      {children}
+    </ul>
+  </>
+);
 
 FolderTree.propTypes = propTypes;
 
-FolderTree.TreeItem = TreeItem;
-
+FolderTree.Item = FolderTreeItem;
 export default FolderTree;