+++ packages/terra-framework-docs/CHANGELOG.md
@@ -3,12 +3,12 @@
 ## Unreleased
 
 * Added
+  * Added examples and tests for `terra-folder-tree`.
   * Added `hasVisibleColumnHeaders` example for FlowsheetDataGrid.
 
 * Updated
   * Updated About page of `terra-table` to provide accessibility documentation for Home and End keys.
   * Removed alpha notice for FlowsheetDataGrid.
-
 
 ## 1.44.0 - (November 9, 2023)